+++ lab1/interface/src/main/java/com/calculator/ICalc.java
@@ -5,4 +5,5 @@ public interface ICalc {
     double subtract(double a, double b);
     double add(double a, double b);
     double divide(double a, double b);
+    double pow(double a, double pow);
 }